Start work on mouse message handling
- Start implementing a label control with a custom window class
- Change WParam/LParam to nint/nuint
- Add some helpers for locking ReaderWriterLockSlim
- Cache window text
- Various cleanup

+++ src/WInterop.Desktop/Gdi/FontHandle.cs
@@ -21,12 +21,16 @@ namespace WInterop.Gdi
             _ownsHandle = ownsHandle;
         }
 
+        public bool IsNull => Handle.IsNull;
+
         public bool IsInvalid => Handle.IsInvalid || Imports.GetObjectType(Handle) != ObjectType.Font;
 
         public void Dispose()
         {
-            if (_ownsHandle)
+            if (_ownsHandle && !Handle.IsInvalid)
+            {
                 Imports.DeleteObject(Handle);
+            }
         }
 
         public static implicit operator HGDIOBJ(FontHandle handle) => handle.Handle;
